lesson-4-agentdeployment/chatkit-server/requirements.txt
# FastAPI server dependencies
fastapi>=0.109.0
uvicorn>=0.27.0

# ChatKit server library
chatkit>=0.1.0

# Azure AI Projects SDK - For calling hosted agents
azure-ai-projects>=1.0.0b1

# Azure Identity - Authentication
azure-identity>=1.15.0

# HTTP client for async requests
httpx>=0.26.0
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0
